+++ NEWS.md
@@ -1,6 +1,7 @@
 # development version 0.2.1.9
 * Fixex bug in get_interval.
 * Removes message that was displayed when incidence class dropped.
+* Refactoring of internal code to improve maintainability.
 
 # incidence2 0.2.1
 * Fixes bug in as.data.frame.incidence2