+++ NEWS.md
@@ -1,5 +1,5 @@
-# development version 0.2.1.9
-* Fixex bug in get_interval.
+# incidence2 0.2.2
+* Fixes bug in get_interval.
 * Removes message that was displayed when incidence class dropped.
 * Refactoring of internal code to improve maintainability.
 * Tests now use the 3rd edition of testthat.